Refactor the SettingsView to have nested pages
