+++ Actuali/Actuali/Views/Settings/ConnectionDataSettingsView.swift
@@ -1,0 +1,449 @@
+import SwiftUI
+
+struct ConnectionDataSettingsView: View {
+    @EnvironmentObject private var budgetStore: BudgetStore
+
+    var body: some View {
+        Form {
+            ServerConnectionSettingsSection()
+
+            // Budget selection needs a live server session. Sync and backups
+            // remain reachable for any loaded budget, including demo and
+            // offline budgets without a session.
+            if budgetStore.isConnected {
+                BudgetSelectionSettingsSection()
+            }
+
+            if budgetStore.currentBudgetId != nil {
+                SyncSettingsSection()
+                BackupSettingsSection()
+            }
+        }
+        .readableWidth()
+        .navigationTitle("Connection & Data")
+        .navigationBarTitleDisplayMode(.inline)
+        .contentMargins(.horizontal, 6, for: .scrollContent)
+    }
+}
+
+private struct ServerConnectionSettingsSection: View {
+    @EnvironmentObject private var budgetStore: BudgetStore
+    @State private var password = ""
+    @State private var showingDisconnectConfirm = false
+
+    var body: some View {
+        Section {
+            TextField("Server URL", text: $budgetStore.serverURL)
+                .textContentType(.URL)
+                .autocapitalization(.none)
+                .keyboardType(.URL)
+                .disabled(budgetStore.isConnected)
+                .accessibilityHint("Example: https://actual.example.com")
+
+            TextField("Fallback server URL (optional)", text: $budgetStore.fallbackServerURL)
+                .textContentType(.URL)
+                .autocapitalization(.none)
+                .keyboardType(.URL)
+                .disabled(budgetStore.isConnected)
+                .accessibilityHint("Used when the primary server cannot be reached")
+
+            if budgetStore.isConnected {
+                HStack {
+                    Text("Status")
+                    Spacer()
+                    HStack(spacing: 4) {
+                        Circle()
+                            .fill(.green)
+                            .frame(width: 8, height: 8)
+                        Text("Connected")
+                            .foregroundStyle(.secondary)
+                    }
+                }
+
+                Button("Disconnect", role: .destructive) {
+                    showingDisconnectConfirm = true
+                }
+            } else {
+                // Show the password field before probing, when password
+                // login is active, or when the first OpenID sign-in needs
+                // the server password (no owner yet).
+                if budgetStore.availableLoginMethods.isEmpty
+                    || budgetStore.passwordLoginActive
+                    || budgetStore.requiresServerPassword {
+                    SecureField(
+                        budgetStore.requiresServerPassword && !budgetStore.passwordLoginActive
+                            ? "Server password (first sign-in)"
+                            : "Password",
+                        text: $password
+                    )
+                }
+
+                Button("Connect") {
+                    Task {
+                        await budgetStore.connect()
+                        guard budgetStore.error == nil else { return }
+                        // Discover available auth methods, then log in with
+                        // password directly when that's the active method.
+                        await budgetStore.checkLoginMethods()
+                        // A probe that couldn't reach the server has
+                        // already explained why; don't repeat the same
+                        // failure as a login attempt.
+                        guard budgetStore.error == nil else { return }
+                        if budgetStore.passwordLoginActive && !password.isEmpty {
+                            await budgetStore.login(password: password)
+                        }
+                    }
+                }
+                .disabled(budgetStore.serverURL.isEmpty || budgetStore.isLoading)
+
+                if budgetStore.supportsOpenIDLogin {
+                    Button("Sign in with OpenID") {
+                        Task {
+                            await budgetStore.loginWithOpenID(
+                                firstTimePassword: password.isEmpty ? nil : password
+                            )
+                        }
+                    }
+                    .disabled(
+                        budgetStore.isLoading
+                            || (budgetStore.requiresServerPassword && password.isEmpty)
+                    )
+                }
+
+                Menu("Try the demo budget") {
+                    Button("Envelope budget") {
+                        Task { await budgetStore.loadDemoData() }
+                    }
+                    Button("Tracking budget") {
+                        Task { await budgetStore.loadDemoData(tracking: true) }
+                    }
+                }
+                .disabled(budgetStore.isLoading)
+
+                NavigationLink {
+                    CustomHeadersEditor(headers: $budgetStore.customHeaders)
+                } label: {
+                    HStack {
+                        Text("Custom HTTP headers")
+                        Spacer()
+                        let count = budgetStore.customHeaders.filter {
+                            !$0.name.trimmingCharacters(in: .whitespaces).isEmpty
+                        }.count
+                        if count > 0 {
+                            Text("\(count)")
+                                .foregroundStyle(.secondary)
+                        }
+                    }
+                }
+            }
+        } header: {
+            Text("Connection")
+        } footer: {
+            if !budgetStore.isConnected {
+                Text("Example: https://actual.example.com\n\nBehind an auth proxy like Cloudflare Access? Add a service token under \u{201C}Custom HTTP headers.\u{201D}\n\nNo server? Tap \u{201C}Try the demo budget\u{201D} to explore the app with sample data. The demo runs entirely on this device \u{2014} it never connects to a server or touches a real budget.")
+            }
+        }
+        .confirmationDialog(
+            "Disconnect and remove data?",
+            isPresented: $showingDisconnectConfirm,
+            titleVisibility: .visible
+        ) {
+            Button("Disconnect & Remove Data", role: .destructive) {
+                budgetStore.logout()
+                password = ""
+            }
+            Button("Cancel", role: .cancel) {}
+        } message: {
+            Text("Signs out and deletes this device's copy of your budgets. Any changes that haven't synced to the server yet will be lost. Your data on the server is not affected.")
+        }
+    }
+}
+
+private struct BudgetSelectionSettingsSection: View {
+    @EnvironmentObject private var budgetStore: BudgetStore
+    @State private var budgetToUnlock: BudgetStore.RemoteBudget?
+    @State private var showingBudgetSelectPrompt = false
+
+    private var selection: Binding<String?> {
+        Binding(
+            get: {
+                // The picker matches on the server fileId, but currentBudgetId
+                // is the internal id — map through the local metadata.
+                guard let budgetId = budgetStore.currentBudgetId,
+                      let metadata = BudgetFileManager.shared.listLocalBudgets().first(
+                        where: { $0.id == budgetId }
+                      ) else {
+                    return nil
+                }
+                return metadata.cloudFileId
+            },
+            set: { newId in
+                if let id = newId,
+                   let budget = budgetStore.remoteBudgets.first(where: { $0.id == id }) {
+                    Task { await budgetStore.downloadBudget(budget) }
+                }
+            }
+        )
+    }
+
+    var body: some View {
+        Section {
+            Picker("Budget", selection: selection) {
+                // Placeholder until a budget is chosen — deliberately
+                // not offered again afterwards, so "None" can't be
+                // (re)selected.
+                if selection.wrappedValue == nil {
+                    Text("Select a Budget").tag(nil as String?)
+                }
+                // Render a placeholder tag for the current cloudFileId
+                // when remoteBudgets hasn't loaded yet (or doesn't
+                // include it), so SwiftUI can match the selection
+                // and we don't get an "invalid selection" warning.
+                if let currentId = selection.wrappedValue,
+                   !budgetStore.remoteBudgets.contains(where: { $0.id == currentId }) {
+                    Text(budgetStore.remoteBudgets.isEmpty ? "Loading…" : "Unknown")
+                        .tag(currentId as String?)
+                }
+                ForEach(budgetStore.remoteBudgets.filter { !$0.isEncrypted }) { budget in
+                    Text(budget.name).tag(budget.id as String?)
+                }
+            }
+            .disabled(budgetStore.downloadingBudgetId != nil)
+
+            ForEach(budgetStore.remoteBudgets.filter(\.isEncrypted)) { budget in
+                Button {
+                    openBudget(budget)
+                } label: {
+                    HStack {
+                        Image(systemName: "lock.fill").foregroundStyle(.secondary)
+                        Text(budget.name)
+                        Spacer()
+                        if budgetStore.downloadingBudgetId == budget.id {
+                            ProgressView()
+                        }
+                    }
+                }
+                .disabled(budgetStore.downloadingBudgetId != nil)
+            }
+
+            if budgetStore.remoteBudgets.isEmpty && !budgetStore.isLoading {
+                Button("Refresh Budgets") {
+                    Task { await budgetStore.fetchRemoteBudgets() }
+                }
+            }
+        } header: {
+            Text("Budget Selection")
+        } footer: {
+            if budgetStore.currentBudgetId == nil {
+                if budgetStore.remoteBudgets.isEmpty && !budgetStore.isLoading {
+                    Text("No budgets were found on your server. Create one in Actual Budget, then tap Refresh Budgets.")
+                } else {
+                    Text("Select a budget to load it onto this device. The app stays empty until one is chosen.")
+                }
+            }
+        }
+        .sheet(item: $budgetToUnlock) { budget in
+            EncryptionPasswordSheet(budget: budget, budgetStore: budgetStore)
+        }
+        .confirmationDialog(
+            "Select a Budget",
+            isPresented: $showingBudgetSelectPrompt,
+            titleVisibility: .visible
+        ) {
+            ForEach(budgetStore.remoteBudgets) { budget in
+                Button(budget.name) {
+                    openBudget(budget)
+                }
+            }
+            Button("Not Now", role: .cancel) {}
+        } message: {
+            Text("You're connected! Choose which budget to load onto this device.")
+        }
+        .task {
+            await budgetStore.fetchRemoteBudgets()
+            promptBudgetSelectionIfNeeded()
+        }
+    }
+
+    private func openBudget(_ budget: BudgetStore.RemoteBudget) {
+        if budget.isEncrypted && EncryptionKeyManager.load(fileId: budget.id) == nil {
+            budgetToUnlock = budget
+        } else {
+            Task { await budgetStore.downloadBudget(budget) }
+        }
+    }
+
+    /// One-time nudge after connecting: surface budget selection so a fresh
+    /// connection doesn't leave the user staring at empty tabs.
+    private func promptBudgetSelectionIfNeeded() {
+        if budgetStore.currentBudgetId == nil && !budgetStore.remoteBudgets.isEmpty {
+            showingBudgetSelectPrompt = true
+        }
+    }
+}
+
+private struct SyncSettingsSection: View {
+    @EnvironmentObject private var budgetStore: BudgetStore
+    @Environment(\.scenePhase) private var scenePhase
+    @State private var lastBackgroundRefresh = BackgroundRefreshStatus().lastRun
+    @State private var refreshRequestError = BackgroundRefreshStatus().lastScheduleError
+    @State private var showingResetSyncConfirm = false
+
+    var body: some View {
+        Section("Sync") {
+            HStack {
+                Text("Status")
+                Spacer()
+                switch budgetStore.syncState {
+                case .idle:
+                    Text("Idle").foregroundStyle(.secondary)
+                case .syncing:
+                    HStack(spacing: 4) {
+                        ProgressView().scaleEffect(0.7)
+                        Text("Syncing...").foregroundStyle(.secondary)
+                    }
+                case .offline:
+                    Text("Offline").foregroundStyle(.orange)
+                case .error:
+                    Text("Error").foregroundStyle(.red)
+                }
+            }
+
+            if case let .error(message) = budgetStore.syncState {
+                Text(message)
+                    .font(.footnote)
+                    .foregroundStyle(.secondary)
+                    .textSelection(.enabled)
+            }
+
+            if budgetStore.syncDetachedByRestore {
+                Text("Sync is disconnected because a backup was restored. Re-download the budget from your server to resume syncing — that replaces the restored data with the server copy.")
+                    .font(.footnote)
+                    .foregroundStyle(.secondary)
+            }
+
+            if let lastSync = budgetStore.lastSyncTime {
+                HStack {
+                    Text("Last Sync")
+                    Spacer()
+                    Text(lastSync, style: .relative)
+                        .foregroundStyle(.secondary)
+                }
+            }
+
+            // Diagnostic: "Never" (or a stale time) with Background
+            // App Refresh enabled usually means the app was
+            // force-quit — iOS won't run the task until next launch.
+            HStack {
+                Text("Last Background Refresh")
+                Spacer()
+                if let lastBackgroundRefresh {
+                    Text(lastBackgroundRefresh, style: .relative)
+                        .foregroundStyle(.secondary)
+                } else {
+                    Text("Never").foregroundStyle(.secondary)
+                }
+            }
+
+            // Distinguishes "the app never asked for a wake" from
+            // "iOS never granted one": a stale row above with no
+            // footnote here points at the system or device
+            // settings, not the app. Hidden while submits succeed
+            // (the common case — we resubmit on every activation).
+            if let refreshRequestError {
+                Text("Refresh request failed: \(refreshRequestError)")
+                    .font(.footnote)
+                    .foregroundStyle(.secondary)
+                    .textSelection(.enabled)
+            }
+
+            Button("Sync Now") {
+                Task { await budgetStore.sync() }
+            }
+            .disabled(budgetStore.syncState == .syncing)
+
+            if case .error = budgetStore.syncState {
+                Button("Reset Sync State", role: .destructive) {
+                    showingResetSyncConfirm = true
+                }
+                .disabled(budgetStore.syncState == .syncing)
+            }
+        }
+        .task { reloadBackgroundRefreshStatus() }
+        // The background refresh task fires while the app is suspended —
+        // same process, so the @State snapshot taken at launch never
+        // re-reads on its own and would show "Never" indefinitely.
+        .onChange(of: scenePhase) { _, newPhase in
+            if newPhase == .active {
+                reloadBackgroundRefreshStatus()
+            }
+        }
+        .confirmationDialog(
+            "Reset sync state?",
+            isPresented: $showingResetSyncConfirm,
+            titleVisibility: .visible
+        ) {
+            Button("Reset & Resync", role: .destructive) {
+                Task { await budgetStore.resetSyncState() }
+            }
+            Button("Cancel", role: .cancel) {}
+        } message: {
+            Text("Discards the local sync marker and re-checks your budget against the server, pulling down anything missing. Local edits are re-sent rather than discarded, but a large budget can take a moment to reconcile.")
+        }
+    }
+
+    private func reloadBackgroundRefreshStatus() {
+        let status = BackgroundRefreshStatus()
+        lastBackgroundRefresh = status.lastRun
+        refreshRequestError = status.lastScheduleError
+    }
+}
+
+private struct BackupSettingsSection: View {
+    @EnvironmentObject private var budgetStore: BudgetStore
+    @State private var confirmBackupOverBaseline = false
+
+    var body: some View {
+        Section {
+            NavigationLink {
+                BackupListView()
+            } label: {
+                HStack {
+                    Text("Backups")
+                    Spacer()
+                    if !budgetStore.backups.isEmpty {
+                        Text("\(budgetStore.backups.count)")
+                            .foregroundStyle(.secondary)
+                    }
+                }
+            }
+
+            Button("Back Up Now") {
+                if budgetStore.isViewingBackup {
+                    confirmBackupOverBaseline = true
+                } else {
+                    Task { await budgetStore.makeBackupNow() }
+                }
+            }
+        } header: {
+            Text("Backups")
+        } footer: {
+            // No cadence promise — iOS can't run a 15-minute
+            // background timer (plan D2).
+            Text("Backups are stored on this device. One is taken automatically when you leave the app.")
+        }
+        .task { await budgetStore.refreshBackups() }
+        .confirmationDialog(
+            "Back up the restored data?",
+            isPresented: $confirmBackupOverBaseline,
+            titleVisibility: .visible
+        ) {
+            Button("Back Up", role: .destructive) {
+                Task { await budgetStore.makeBackupNow() }
+            }
+            Button("Cancel", role: .cancel) {}
+        } message: {
+            Text("This makes the restored data your current budget and removes the option to revert to the version from before you loaded a backup.")
+        }
+    }
+}
