+++ Actuali/Actuali/Views/Settings/WalletAutomationView.swift
